RGB to HSBK conversion function
Useful as many libraries and use-cases only deal with the more common RGB colour format (whereas LifX requires HSBK). For example: using 'Pillow' to analyse a screenshot for an ambi-light esque  application.

Given it is not directly linked to any of the other files I propose creating a generic 'accessory' file where any extraneous but useful functions for LifX users can be placed.